Serial Implementation

+++ src/filemodule.cc
@@ -2,11 +2,11 @@
 
 int file_road_size(std::string filename){
     int num_values = 0;
-    int target;
+    double target;
     FILE *curr_file;
     curr_file = fopen(filename.c_str(), "r");
 
-    for (; fscanf(curr_file, "%d", &target) != EOF; num_values++){}
+    for (; fscanf(curr_file, "%lf", &target) != EOF; num_values++){}
     
     fclose(curr_file);
     return num_values;
